requirements.txt
html==1.16
image==1.5.27
Pillow==6.2.1
PyGithub==1.44.1
